Remove unnecessary long to double conversion
As there were only two different overloaded methods that had the int and the double as an input, java decided that when the method is getting accessed by a double it converts it into a double.

But this is something that is not intended as doubles are not precise which can be crucial at some points in the application.