Change the generated cass to an enum. (#2)
* Change the generated class to an enum.

Changing the processor so that it now uses the Enum File generator.

Moving the Java class generators into their own package.
Add another generator that is based on the JClassGenerator that generates enums.

* Remove unnecessary imports.

The imports of Map and HashMap were used in the previous version where a class was used but as now an enum is being used these are not useful anymore.

* Get the names of the classes from the annotation

Before the names of the classes were implied which meant that no class could be renamed that was used by this generated class. But now the generated class is changing with the refactored class.